v14.12
model with hiperpriors per group, i.e., partial pooling.

the diagnostics are better, although the hiperpriors for forest structure are still not very good.

I now also plot the trajectories in relation to the observed data. this shows that the prior for old-growth is too permissive in terms of the variance. this will be changed for the next version